+++ Changelog.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Move the startupinit module from atomic/atom to utils. Completed on 2026-01-01
+
 ### Deprecated
 
 - No longer to use guimaker from atomic/atom. Completed on 2026-01-01